separate userdefined list into separate module
